extract db initialization from `instrumentation.ts` into `/instrumentation/`
